requirements.txt
streamlit>=1.35.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
accelerate>=0.29.0
scikit-learn>=1.4.0
xgboost>=2.0.0
mlflow>=2.12.0
Pillow>=10.0.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
umap-learn>=0.5.6
